Complete numeric literal and mixed-width arithmetic change.
- Contextual numeric literal typing for all integer/float widths.
- Mixed-width arithmetic and promotion to a smallest common type.
- Lossy Float128 to-string via f64 truncation so it can print.
- LLVM codegen fixes (NonZeroU32, Shl over Float, ConstInt).
- zirk-sema, zirk-ir, zirk-codegen-llvm, and CLI corpus tests.
- Archive OpenSpec change and sync main specs.

+++ crates/zirk-ir/src/ir.rs
@@ -790,7 +790,7 @@ impl Instruction {
 /// What an instruction does.
 #[derive(Debug, Clone, PartialEq)]
 pub enum InstKind {
-    ConstInt(i32),
+    ConstInt(i128),
     ConstBool(bool),
     /// Materializes a string literal. This is the only allocating operation of
     /// the subset, and it is expressed without naming a memory strategy.
